+++ packages/flutter_chat_view/lib/src/screens/chat_screen.dart
@@ -384,7 +384,6 @@ class ChatListItem extends StatelessWidget {
               ),
             ),
           ),
-          const Divider(),
         ],
       );
 }